Deleted WorkPlace mapper based on mapstruct; added EmployeeWorkPlace entity and "active" field in it

+++ src/main/java/com/example/javaadv_task_5/service/WorkPlaceService.java
@@ -4,5 +4,5 @@ import com.example.javaadv_task_5.domain.WorkPlace;
 
 public interface WorkPlaceService {
   WorkPlace create(WorkPlace workPlace);
-  WorkPlace getById(Integer id);
+  WorkPlace getById(Long id);
 }
